# open_api_docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- {}
paths:
...
  /api/fine-tuning/{suggestedQuestion}/duplicate:
    post:
      tags:
      - fine-tuning-data-controller
      operationId: duplicate
      parameters:
      - name: suggestedQuestion
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SuggestedQuestionForFineTuning"
  /api/fine-tuning/{suggestedQuestion}/delete:
    post:
      tags:
      - fine-tuning-data-controller
      operationId: delete
      parameters:
      - name: suggestedQuestion
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SuggestedQuestionForFineTuning"
  /api/fine-tuning/upload-and-trigger-fine-tuning:
    post:
      tags:
      - fine-tuning-data-controller
      operationId: uploadAndTriggerFineTuning
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
  /api/conversation/{conversationId}/send:
    post:
      tags:
      - conversation-message-controller
      operationId: replyToConversation
      parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConversationMessage"
  /api/conversation/{conversationId}/ai-reply:
    post:
      tags:
      - conversation-message-controller
      operationId: getAiReply
      parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SseEmitter"
  /api/conversation/recall-prompt/{recallPrompt}:
    post:
      tags:
      - conversation-message-controller
      operationId: startConversationAboutRecallPrompt
      parameters:
      - name: recallPrompt
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Conversation"
  /api/conversation/note/{note}:
    get:
      tags:
      - conversation-message-controller
      operationId: getConversationsAboutNote
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      tags:
      - conversation-message-controller
      operationId: startConversationAboutNote
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Conversation"
  /api/conversation/assessment-question/{assessmentQuestion}:
    post:
      tags:
      - conversation-message-controller
      operationId: startConversationAboutAssessmentQuestion
      parameters:
      - name: assessmentQuestion
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Conversation"
...
  /api/fine-tuning/{suggestedQuestion}/update-suggested-question-for-fine-tuning:
    patch:
      tags:
      - fine-tuning-data-controller
      operationId: updateSuggestedQuestionForFineTuning
      parameters:
      - name: suggestedQuestion
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionSuggestionParams"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SuggestedQuestionForFineTuning"
  /api/conversation/{conversationId}/read:
    patch:
      tags:
      - conversation-message-controller
      operationId: markConversationAsRead
      parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationMessage"
...
  /api/fine-tuning/all-suggested-questions-for-fine-tuning:
    get:
      tags:
      - fine-tuning-data-controller
      operationId: getAllSuggestedQuestions
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuggestedQuestionForFineTuning"
...
  /api/conversation/{conversationId}:
    get:
      tags:
      - conversation-message-controller
      operationId: getConversation
      parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Conversation"
  /api/conversation/{conversationId}/messages:
    get:
      tags:
      - conversation-message-controller
      operationId: getConversationMessages
      parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationMessage"
  /api/conversation/{conversationId}/export:
    get:
      tags:
      - conversation-message-controller
      operationId: exportConversation
      parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /api/conversation/unread:
    get:
      tags:
      - conversation-message-controller
      operationId: getUnreadConversations
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationMessage"
  /api/conversation/all:
    get:
      tags:
      - conversation-message-controller
      operationId: getConversationsOfCurrentUser
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
...
components:
  schemas:
    Circle:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          maxLength: 255
          minLength: 1
      required:
      - id
      - name
    Ownership:
      type: object
      properties:
        id:
          type: integer
          format: int32
        circle:
          $ref: "#/components/schemas/Circle"
        ownerName:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        externalIdentifier:
          type: string
        ownership:
          $ref: "#/components/schemas/Ownership"
        dailyAssimilationCount:
          type: integer
          format: int32
        spaceIntervals:
          type: string
        admin:
          type: boolean
      required:
      - externalIdentifier
      - id
      - name
...
    MCQWithAnswer:
      type: object
      properties:
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
        correctChoiceIndex:
          type: integer
          format: int32
        strictChoiceOrder:
          type: boolean
      required:
      - correctChoiceIndex
      - multipleChoicesQuestion
      - strictChoiceOrder
    MultipleChoicesQuestion:
      type: object
      properties:
        stem:
          type: string
        choices:
          type: array
          items:
            type: string
      required:
      - choices
      - stem
    QuestionSuggestionParams:
      type: object
      properties:
        comment:
          type: string
        preservedQuestion:
          $ref: "#/components/schemas/MCQWithAnswer"
        preservedNoteContent:
          type: string
        positiveFeedback:
          type: boolean
        realCorrectAnswers:
          type: string
      required:
      - preservedQuestion
      - realCorrectAnswers
...
    PredefinedQuestion:
      type: object
      properties:
        id:
          type: integer
          format: int32
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
        correctAnswerIndex:
          type: integer
          format: int32
        approved:
          type: boolean
      required:
      - id
      - multipleChoicesQuestion
...
    Note:
      type: object
      properties:
        noteTopology:
          $ref: "#/components/schemas/NoteTopology"
        details:
          type: string
        parentId:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          readOnly: true
        wikidataId:
          type: string
      required:
      - createdAt
      - id
      - noteTopology
      - updatedAt
    NoteTopology:
      type: object
      properties:
        id:
          type: integer
          format: int32
        titleOrPredicate:
          type: string
        shortDetails:
          type: string
        linkType:
          type: string
          enum:
          - no link
          - related to
          - a specialization of
          - an application of
          - an instance of
          - a part of
          - tagged by
          - an attribute of
          - the opposite of
          - author of
          - using
          - an example of
          - before
          - similar to
          - confused with
        objectNoteTopology:
          $ref: "#/components/schemas/NoteTopology"
        parentOrSubjectNoteTopology:
          $ref: "#/components/schemas/NoteTopology"
        title:
          type: string
      required:
      - id
      - titleOrPredicate
...
    Answer:
      type: object
      properties:
        id:
          type: integer
          format: int32
        choiceIndex:
          type: integer
          format: int32
        correct:
          type: boolean
      required:
      - correct
      - id
    AnsweredQuestion:
      type: object
      properties:
        note:
          $ref: "#/components/schemas/Note"
        predefinedQuestion:
          $ref: "#/components/schemas/PredefinedQuestion"
        answer:
          $ref: "#/components/schemas/Answer"
        answerDisplay:
          type: string
        recallPromptId:
          type: integer
          format: int32
      required:
      - answer
      - predefinedQuestion
      - recallPromptId
...
    SuggestedQuestionForFineTuning:
      type: object
      properties:
        id:
          type: integer
          format: int32
        comment:
          type: string
        preservedQuestion:
          $ref: "#/components/schemas/MCQWithAnswer"
        preservedNoteContent:
          type: string
        realCorrectAnswers:
          type: string
        createdAt:
          type: string
          format: date-time
        positiveFeedback:
          type: boolean
      required:
      - id
      - preservedQuestion
...
    ConversationMessage:
      type: object
      properties:
        id:
          type: integer
          format: int32
        message:
          type: string
        sender:
          $ref: "#/components/schemas/User"
        readByReceiver:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
      - id
      - message
    SseEmitter:
      type: object
      properties:
        timeout:
          type: integer
          format: int64
    AssessmentQuestionInstance:
      type: object
      properties:
        id:
          type: integer
          format: int32
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
        answer:
          $ref: "#/components/schemas/Answer"
      required:
      - id
      - multipleChoicesQuestion
    Conversation:
      type: object
      properties:
        id:
          type: integer
          format: int32
        subject:
          $ref: "#/components/schemas/ConversationSubject"
        subjectOwnership:
          $ref: "#/components/schemas/Ownership"
        conversationInitiator:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - createdAt
      - id
      - updatedAt
    ConversationSubject:
      type: object
      properties:
        assessmentQuestionInstance:
          $ref: "#/components/schemas/AssessmentQuestionInstance"
        note:
          $ref: "#/components/schemas/Note"
        answeredQuestion:
          $ref: "#/components/schemas/AnsweredQuestion"
...